agents/smolagent/pyproject.toml
[project]
name = "smolagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.5.5",
    "huggingface-hub>=0.34.4",
    "langfuse>=3.3.4",
    "openinference-instrumentation-smolagents>=0.1.16",
    "opentelemetry-exporter-otlp>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "smolagents>=1.21.3",
]
